// packages/air-ui-ds/models/enums/progress.ts
export enum ProgressBarSize {
    XS = 'xs',
    SM = 'sm',
    MD = 'md',
    LG = 'lg',
    XL = 'xl'
}

export enum ProgressCircleSize {
    XS = 'xs',
    SM = 'sm',
    MD = 'md',
    LG = 'lg',
    XL = 'xl',
    XXL = '2xl',
}

export enum ProgressCircleLabelPosition {
    CENTER = 'center',
    TOP = 'top',
    BOTTOM = 'bottom',
}